[shuffle] move e2e tests to top lvl project folder
