+++ shuffle/move/examples/e2e/utils.ts
@@ -1,7 +1,7 @@
 // Copyright (c) The Diem Core Contributors
 // SPDX-License-Identifier: Apache-2.0
 
-export * as TxnBuilder from "../txn_builders/mod.ts";
+export * as TxnBuilder from "../Message/txn_builders/mod.ts";
 
 export function deployMessageModule() {
   return true;
